Combine `utils/helpers.ts` and `utils/commonHelpers.ts`
